Check for RWD_DATA environment variable
Previously, we had always assumed that developers would have a folder on
their workstation containing RWD NHD data. However, the data is so large
that we need to use external hard drives. Because of this, symlinks were
used to point to the mounted volume location. Setting all of this up,
especially for developers that have not worked on this feature, can be
confusing. Using an environment variable gives us the ability to point
to the correct location without having to rely on symlinks or hacking
the Vagrantfile.

I have also decided against providing a default value to avoid making
assumptions about where people store this data.